Now able to disable DB purging

+++ db.ts
@@ -6,11 +6,18 @@ import {Ranks}				from "./classes";
 
 // Delete old table in debug, to get rid of old data.
 // DISABLE IN PRODUCTION!
-try {
-	fs.unlinkSync("./db/db.sqlite3");
-} catch (e) {
-	console.error(e);
+
+const PURGE_DATABASE = false;
+
+if (PURGE_DATABASE) {
+	try {
+		fs.unlinkSync("./db/db.sqlite3");
+	} catch (e) {
+		console.error(e);
+	}
 }
+
+
 
 
 export const db = bsqlite3("./db/db.sqlite3");
@@ -19,60 +26,68 @@ export const db = bsqlite3("./db/db.sqlite3");
 db.pragma("journal_mode = WAL");
 
 
-// Make tables
-db.prepare(`
-	CREATE TABLE IF NOT EXISTS users(
-		name			TEXT		NOT NULL,
-		uuid			TEXT		NOT NULL,
-		passHash		TEXT		NOT	NULL,
-		date			DATETIME	NOT NULL DEFAULT CURRENT_TIMESTAMP,
-		adminRank		INTEGER		NOT NULL DEFAULT 0,
-		emailVerified	BOOLEAN		NOT NULL DEFAULT 0
-						CHECK (emailVerified IN (0, 1)),
-		emailVToken		TEXT,
-		authToken		TEXT,
-		bio				TEXT
-	);
-`).run();
 
-db.prepare(`
-	CREATE TABLE IF NOT EXISTS news(
-		title		TEXT		NOT NULL,
-		author		TEXT		NOT NULL DEFAULT 'Anonymous',
-		content		TEXT		NOT NULL,
-		visible		BOOLEAN		NOT NULL DEFAULT 1 CHECK (visible IN (0, 1)),
-		date		DATETIME	NOT NULL DEFAULT CURRENT_TIMESTAMP
-	);
-`).run();
 
-db.prepare(`
-	CREATE TABLE IF NOT EXISTS games(
-		id			TEXT		PRIMARY KEY,
-		title		TEXT		NOT NULL,
-		creator		TEXT		NOT NULL DEFAULT 'Anonymous',
-		description	TEXT		NOT NULL DEFAULT 'No description given... :(',
-		visible		BOOLEAN		NOT NULL DEFAULT 1 CHECK (visible IN (0, 1)),
-		date		DATETIME	NOT NULL DEFAULT CURRENT_TIMESTAMP
-	);
-`).run();
-
-db.prepare(`
-	CREATE TABLE IF NOT EXISTS capsules(
-		uuid		TEXT		PRIMARY KEY,
-		name		TEXT		NOT NULL,
-		creator		TEXT		NOT NULL,
-		version		TEXT		NOT NULL,
-		content		TEXT		NOT NULL,
-		visible		BOOLEAN		NOT NULL DEFAULT 1 CHECK (visible IN (0, 1)),
-		date		DATETIME	NOT NULL DEFAULT CURRENT_TIMESTAMP,
-		likes		INTEGER		NOT NULL DEFAULT 0
-	);
-`).run();
+if (PURGE_DATABASE) {
+	// Make tables
+	db.prepare(`
+		CREATE TABLE IF NOT EXISTS users(
+			name			TEXT		NOT NULL,
+			uuid			TEXT		NOT NULL,
+			passHash		TEXT		NOT	NULL,
+			date			DATETIME	NOT NULL DEFAULT CURRENT_TIMESTAMP,
+			adminRank		INTEGER		NOT NULL DEFAULT 0,
+			emailVerified	BOOLEAN		NOT NULL DEFAULT 0
+							CHECK (emailVerified IN (0, 1)),
+			emailVToken		TEXT,
+			authToken		TEXT,
+			bio				TEXT
+		);
+	`).run();
+	
+	db.prepare(`
+		CREATE TABLE IF NOT EXISTS news(
+			title		TEXT		NOT NULL,
+			author		TEXT		NOT NULL DEFAULT 'Anonymous',
+			content		TEXT		NOT NULL,
+			visible		BOOLEAN		NOT NULL DEFAULT 1 CHECK (visible IN (0, 1)),
+			date		DATETIME	NOT NULL DEFAULT CURRENT_TIMESTAMP
+		);
+	`).run();
+	
+	db.prepare(`
+		CREATE TABLE IF NOT EXISTS games(
+			id			TEXT		PRIMARY KEY,
+			title		TEXT		NOT NULL,
+			creator		TEXT		NOT NULL DEFAULT 'Anonymous',
+			description	TEXT		NOT NULL DEFAULT 'No description given... :(',
+			visible		BOOLEAN		NOT NULL DEFAULT 1 CHECK (visible IN (0, 1)),
+			date		DATETIME	NOT NULL DEFAULT CURRENT_TIMESTAMP
+		);
+	`).run();
+	
+	db.prepare(`
+		CREATE TABLE IF NOT EXISTS capsules(
+			uuid		TEXT		PRIMARY KEY,
+			name		TEXT		NOT NULL,
+			creator		TEXT		NOT NULL,
+			version		TEXT		NOT NULL,
+			content		TEXT		NOT NULL,
+			visible		BOOLEAN		NOT NULL DEFAULT 1 CHECK (visible IN (0, 1)),
+			date		DATETIME	NOT NULL DEFAULT CURRENT_TIMESTAMP,
+			likes		INTEGER		NOT NULL DEFAULT 0
+		);
+	`).run();
+}
 
 
 
 import statements			from "./statements";
-statements.prepopulate.forEach(v => v.run());
+if (PURGE_DATABASE) {
+	statements.prepopulate.forEach(v => v.run());
+}
+
+
 
 export namespace accs {
 	export async function register(user: string, pass: string) {
@@ -98,6 +113,8 @@ export namespace accs {
 		statements.updateBio.run(bio, user);
 	}
 }
+
+
 
 [//	Reserved Username	// Reason
 	"Dexie",			// Admin account
